Fix shadow compute not using all LOD slices

+++ crest/Assets/Crest/Crest/Shaders/OceanInputsDriven.hlsl
@@ -25,7 +25,7 @@ float4 _GeomData;
 // _LD_Params: float4(world texel size, texture resolution, shape weight multiplier, 1 / texture resolution)
 float4 _LD_Params[MAX_LOD_COUNT + 1];
 float3 _LD_Pos_Scale[MAX_LOD_COUNT + 1];
-float _LD_SliceIndex;
+uint _LD_SliceIndex;
 float4 _LD_Params_Source[MAX_LOD_COUNT + 1];
 float3 _LD_Pos_Scale_Source[MAX_LOD_COUNT + 1];
 CBUFFER_END
